refactor: invert engine.diff -> api dependency; add architecture guard test
engine/diff lazily imported sdpm.api (generate, template resolution) to
load/build diff inputs — a dependency-rule violation (core importing the
facade), inherited from main and flagged in the v0.5 review.

- engine.diff is now pure comparison: diff_slides(base: dict, edit: dict)
- loading/building (deck dirs, source JSONs, PPTX roundtrips) moved to
  sdpm.api.load_slides_json_or_pptx + sdpm.api.diff_report (facade
  orchestration, same code path); callers (tools contract, CLI, tests)
  updated
- tests/test_architecture.py: guards the dependency rule — engine/ and
  knowledge/ must never import sdpm.api or sdpm.tools

SPEC: 20260731-0820_v0.5.x-consolidation-roadmap (architecture debt)
